+++ script.lua
@@ -2,12 +2,18 @@ local s = sonopix
 
 local opened = s.open_file("/home/neo/Downloads/abstract-whale-png.png")
 if opened then
+    print(s.opts.cursor.color)
     s.opts = {
         direction = "top-to-bottom",
         frequency_range = { 20, 20000 },
-        spu = 0.1,
+        cursor = {
+            width = 5,
+            color = "#FF5000FF",
+        },
+        -- spu = 0.1,
         amplitude = 0.5,
     }
+    -- s.opts.direction = "top-to-bottom"
 
     if s.sonify() then
         s.play()